+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 # UNRELEASED
+
+**New Features:**
+- add new `touchEventOptions` prop that can set the options for the touch event listeners
+  - this provides users full control of if/when they want to set [passive](https://developer.mozilla.org/en-US/docs/Web/API/EventTarget/addEventListener#options)
+- add new `onTouchStartOrOnMouseDown` prop that is called for `touchstart` and `mousedown`. Before a swipe even starts.
+  - combined with `touchEventOptions` allows users the ability to now call `preventDefault` on `touchstart`
+- add new `onTouchEndOrOnMouseUp` prop that is called for `touchend` and `mouseup`.
+
 **Breaking Changes:**
 * we have dropped support for `es5` transpiled output
   * we target `es2015` for our transpilation now
@@ -6,9 +14,10 @@
 * `preventScrollOnSwipe` - "new" prop. Replaces `preventDefaultTouchmoveEvent`
   * same functionality but renamed to be more explicit on its intended use
   * **fixed bug** - where toggling this prop did not re-attach event listeners
-  * **small update** - we now only change the `passive` event listener option for `touchmove` depending on this prop
+  * **update** - we now **only** change the `passive` event listener option for `touchmove` depending on this prop
     * see notes in README for more details [readme#passive-listener](https://github.com/FormidableLabs/react-swipeable#passive-listener)
   * Thank you [@stefvhuynh](https://github.com/stefvhuynh)
+
 **Bug fixes:**
 * fix bug where directional swiped check allowed `undefined`/falsy values to set `cancelablePageSwipe`
   * Thank you [@bhj](https://github.com/bhj) for the [comment](https://github.com/FormidableLabs/react-swipeable/pull/240#issuecomment-1014980025)